requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
requests>=2.31.0
pydantic>=2.5.0
pydantic-settings>=2.1.0 